fix(rc.d): Don't stomp on daemon's pid
Previous behavior would overwrite the pid of the daemon process, thus rc.d is killing the wrong process.

Ticket: #29185

+++ src/middlewared/middlewared/main.py
@@ -1158,8 +1158,9 @@ def main():
     os.environ['MIDDLEWARED'] = str(os.getpid())
 
     if args.foreground:
-        with open(pidpath, "w") as _pidfile:
-            _pidfile.write(f"{str(os.getpid())}\n")
+        if not os.path.exists(pidpath):
+            with open(pidpath, "w") as _pidfile:
+                _pidfile.write(f"{str(os.getpid())}\n")
 
     Middleware(
         loop_monitor=not args.disable_loop_monitor,
